Add support for MRL B devices.
This patch adds support for MRL B devices by
masking the right most nibble when the chip
is identified at start of main loop.

MRL B devices are identified by having a 1 in
the right most nibble of CHIPID_CIDR.

Without masking that nibble during compare,
usamba fails with an error message when an MRL B
device is connected.

Signed-off-by: Patrick Menschel <menschel.p@posteo.de>